requirements.txt
streamlit
pymupdf
pymongo
sentence-transformers==2.7.0
faiss-cpu
pandas
nltk
scikit-learn